refactor: improve login error handling, env type checks, toast root, merge schema
