+++ .env.example
@@ -1,0 +1,6 @@
+VITE_CTP_PROJECT_KEY=your-project-key
+VITE_CTP_CLIENT_SECRET=your-client-secret
+VITE_CTP_CLIENT_ID=your-client-id
+VITE_CTP_AUTH_URL=auth-url
+VITE_CTP_API_URL=api-url
+VITE_CTP_SCOPES=scopes
