+++ src/lib/schemas/user.ts
@@ -1,5 +1,4 @@
 import { z } from 'zod';
-import { loginSchema } from './login-schema';
 
 const patterns: Record<string, [RegExp, string]> = {
   NO_SPECIAL_CHARS: [/^[^!@#$%^&*]*$/, 'Must contain no special characters'],
@@ -13,17 +12,42 @@ const patterns: Record<string, [RegExp, string]> = {
 
 const MIN_AGE = 13;
 
+export const loginSchema = z.object({
+  email: z
+    .string()
+    .min(1, 'Required field')
+    .regex(/^(\S|$)/, 'Must not contain leading whitespace(s)')
+    .regex(/.*\S$|^$/, 'Must not contain trailing whitespace(s)')
+    .regex(/\b(?:[a-z0-9-]+\.)+[a-z]{2,}\b/, 'Must contain a domain name (e.g., example.com)')
+    .regex(/^[^@]+@[^@]+$/, 'Must contain a separating "@" symbol')
+    .email('Must be properly formatted (e.g., user@example.com)'),
+
+  password: z
+    .string()
+    .min(8, 'At least 8 characters long')
+    .regex(/[A-Z]/g, 'At least one uppercase letter (A-Z)')
+    .regex(/[a-z]/g, 'At least one lowercase letter (a-z)')
+    .regex(/[0-9]/g, 'At least one digit (0-9)')
+    .regex(/[!@#$%^&*]/g, 'At least one special character (e.g., !@#$%^&*)')
+    .regex(/^(\S|$)/, 'Must not contain leading whitespace(s)')
+    .regex(/.*\S$|^$/, 'Must not contain trailing whitespace(s)'),
+});
+
+export type LoginSchema = z.infer<typeof loginSchema>;
+
 const baseSchema = loginSchema.extend({
   firstName: z
     .string()
     .min(1, 'Required field')
     .regex(...patterns.NO_SPECIAL_CHARS)
     .regex(...patterns.NO_NUMBERS),
+
   lastName: z
     .string()
     .min(1, 'Required field')
     .regex(...patterns.NO_SPECIAL_CHARS)
     .regex(...patterns.NO_NUMBERS),
+
   dateOfBirth: z.date().refine((userDate) => {
     const today = new Date();
     today.setHours(0, 0, 0, 0);
@@ -36,6 +60,7 @@ const baseSchema = loginSchema.extend({
   }, `You must be at least ${MIN_AGE} years old`),
 
   street: z.string().min(1, 'Required field'),
+
   city: z
     .string()
     .min(1, 'Required field')
